svc: Add an implicit 'l' option to start the output to supervise
If a script issued 'svc -L...' on a service, particularly with the -u
option, and then immediately issued another 'svc -...' without either -L
or -l option, supervise may read the second commands as part of the
first, and issue them to the log service instead of the main one.  By
implicitly adding an 'l' option to the data sent to supervise, this flaw
is avoided.