Find FIONREAD includes for the fifo program:
Linux, BSD, OS/X, and others seem to favor sys/ioctl.h. Some Sun machines
like sys/filio.h, both with and without sys/ioctl.h. At least Minix uses
sys/ioc_file.h. On top of that, some libraries define the ioctl with a
pointer to int while others point to long.

Added compile-and-run test cases to cover all the above. Should be fairly
painless to add if the need arises.

If FIONREAD is not found however, that bit of code is ifdef'ed, so
shouldn't be a real problem. The drawback in that case is that the fifo
program could start reading before its child starts. If the child fails
to start, fifo dies and whatever has been read disappears with it.