don't remove status if locked
if there was a lock file, supervise correctly didn't start, but removed
the status file of the other supervise that was most likely running at
the same time, why else would there be a lock?

This fixes that. Now the file is truncated just before the first
announce. Technically one wouldn't want to truncate at all, but that's
not so easy.